# swagger.yaml
swagger: "2.0"
info:
  description: |-
    The purpose of this application is to manage collections of todos owned by
    authenticated users
  version: 1.0.0
  title: Todo API
host: localhost
basePath: /
schemes:
- http
...
paths:
...
  /todos:
    get:
      summary: Get all todos for the current User
      description: |-
        Returns a list of all todos belonging to the current authenticated user.
      operationId: GetTodos
      produces:
      - application/json
      tags:
      - Todo
      responses:
        "200":
          description: "list of todos"
          schema:
            type: array
            items:
              $ref: '#/definitions/schemas/Todo'
    post:
      summary: Add a single Todo
      description: Adds a todo belonging to the current authenticated user
      operationId: AddTodo
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - Todo
      parameters:
      - in: "body"
        name: "body"
        description: "Details of the Todo to be added"
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string"
              description: A short name or description of the Todo
              example: "Do some things"
            completed:
              type: "string"
              format: date-time
              description: The time the Todo was completed. Pass null or exclude parameter if the Todo is incomplete
          required:
            - name
      responses:
        "201":
          description: "todo that was created"
          schema:
            $ref: '#/definitions/schemas/Todo'
...